http_header: move all code into keyword files
