+++ src/detect-http-header.h
@@ -27,4 +27,6 @@
 void DetectHttpHeaderRegister(void);
 void DetectHttpRawHeaderRegister(void);
 
+void DetectEngineCleanHHDBuffers(DetectEngineThreadCtx *det_ctx);
+
 #endif /* __DETECT_HTTP_HEADER_H__ */
